+++ livekit-plugins/install_plugins_editable.sh
@@ -5,7 +5,8 @@ if [[ -z "$VIRTUAL_ENV" ]]; then
     echo "You are not in a virtual environment."
     exit 1
 fi
-
+pip install -e ./livekit-plugins-anthropic --config-settings editable_mode=strict
+pip install -e ./livekit-plugins-aws --config-settings editable_mode=strict
 pip install -e ./livekit-plugins-assemblyai --config-settings editable_mode=strict
 pip install -e ./livekit-plugins-azure --config-settings editable_mode=strict
 pip install -e ./livekit-plugins-cartesia --config-settings editable_mode=strict
@@ -21,5 +22,6 @@ pip install -e ./livekit-plugins-rime --config-settings editable_mode=strict
 pip install -e ./livekit-plugins-llama-index --config-settings editable_mode=strict
 pip install -e ./livekit-plugins-turn-detector --config-settings editable_mode=strict
 pip install -e ./livekit-plugins-silero --config-settings editable_mode=strict
+pip install -e ./livekit-plugins-speechmatics --config-settings editable_mode=strict
 pip install -e ./livekit-plugins-browser --config-settings editable_mode=strict