+++ livekit-plugins/livekit-plugins-nltk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # livekit-plugins-nltk
 
+## 0.7.4
+
+### Patch Changes
+
+- updated livekit-agent reference to <1.0 - [#1607](https://github.com/livekit/agents/pull/1607) ([@davidzhao](https://github.com/davidzhao))
+
 ## 0.7.3
 
 ### Patch Changes