+++ livekit-plugins/livekit-plugins-playai/CHANGELOG.md
@@ -1,5 +1,19 @@
 # livekit-plugins-playht
 
+## 1.0.9
+
+### Patch Changes
+
+- use streaming AudioDecoder to handle compressed encoding - [#1584](https://github.com/livekit/agents/pull/1584) ([@davidzhao](https://github.com/davidzhao))
+
+- updated livekit-agent reference to <1.0 - [#1607](https://github.com/livekit/agents/pull/1607) ([@davidzhao](https://github.com/davidzhao))
+
+## 1.0.8
+
+### Patch Changes
+
+- remove update options from tts synthesis stream - [#1546](https://github.com/livekit/agents/pull/1546) ([@jayeshp19](https://github.com/jayeshp19))
+
 ## 1.0.7
 
 ### Patch Changes